src/main/java/com/javabrains/moviecatalogservice/resources/MovieCatalogResource.java
package com.javabrains.moviecatalogservice.resources;

import com.javabrains.moviecatalogservice.models.CatalogItem;
import com.javabrains.moviecatalogservice.models.Movie;
import com.javabrains.moviecatalogservice.models.Rating;
import com.javabrains.moviecatalogservice.models.UserRating;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.client.RestTemplate;
import org.springframework.web.reactive.function.client.WebClient;
import java.util.Arrays;
import java.util.Collections;
import java.util.List;
import java.util.stream.Collectors;

@RestController
@RequestMapping("/catalog")
public class MovieCatalogResource {

    @Autowired
    private RestTemplate restTemplate;

    @Autowired
    private WebClient.Builder webClientBuilder;

    @RequestMapping("/{userId}")
    public List<CatalogItem> getCatalog(@PathVariable("userId") String userId){
        //RestTemplate restTemplate = new RestTemplate();
        //get all rated movie ids Hard coded now

        UserRating ratings = restTemplate.getForObject("http://ratings-data-service/ratingsdata/users/"+userId, UserRating.class);
        return ratings.getUserRating().stream().map(rating -> {
            //for each movie id, call movie info service and get details
            Movie movie = restTemplate.getForObject("http://movie-info-service/movies/" + rating.getMovieId(), Movie.class);
            //put them all together
            return new CatalogItem(movie.getName(),"A rare and true love story",rating.getRating());

        }).collect(Collectors.toList());

    }
}

//        return Collections.singletonList(
//                new CatalogItem("Sita Raman","A rare and true love story",5)
//        );

//            Movie movie = webClientBuilder.build()
//                    .get()
//                    .uri("http://localhost:8082/movies/" + rating.getMovieId())
//                    .retrieve()
//                    .bodyToMono(Movie.class) //returning asynchronous obj
//                    .block();
